requirement.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
pillow
streamlit
requests
tqdm